requirements/prod.txt
uwsgi
django-redis-sessions
